Implement reading side of subroutines.
Whenever a program is created, it queries all the information needed
about its subroutines (see new code in reflection.rs).

The next step in implementing subroutines will be to create an interface
so active subroutines can be customized when calling `draw`. The
information needed to validate the submitted data now exists.