Implement subroutines.
This commit adds functionality to use subroutines as uniform values.
To do this, the user has to give a tuple `("subroutine_name", ShaderStage)` to the `uniform!` macro.
The implementation makes sure that all subroutines of a certain stage
must be present and that the name strings are actually valid
subroutines.

CC #405 #1146