Adjust compute shader example
This addresses issue #1310 where the compute shader example,
gpgpus.rs, does not run on certain setups.  With the most
recent up to date nVidia driver the example would have no issues
but on an older driver, capable of running OpenGL 4.5 and over a
year old, the example would panic due to a compilation error of
the shader.