Silence unused imports warning in examples
glium reexports glutin and I think this
point needs to be stressed. So using glutin
from the reexport instead of the possibly
available extern crate is useful at least
in the examples otherwise users might not
realize that they don't have to add glutin
to their Cargo.toml.

Therefore, the lint is silenced instead of
being followed.

Link to the lint change that likely caused this:

https://github.com/rust-lang/rust/pull/58805