Make `Content::read()` an unsafe API
Two provided impls of `Content::read()` are unsafe in that
they can either (1)expose a zeroed `T` that cannot be `T` or (2) expose
an uninitialized vector of `T` to a user-provided function.

Currently, the `Content::read()` API can be called from safe Rust so
that it is possible to trigger undefined behavior using safe Rust APIs.

I learned from discussion in #1907 that user-provided function is not
expected to read but only expected to write to `&mut Self`, but I
couldn't find this description in the documentation.

This commit switched `Content::read()` API to an unsafe API, and also
added warning message to documentation.